requirements.txt
anyio==4.8.0
certifi==2024.12.14
googletrans==4.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.0.1
idna==3.10
packaging==24.2
pillow==11.1.0
pypinyin==0.53.0
pytesseract==0.3.13
sniffio==1.3.1
typing_extensions==4.12.2
